+++ lib/examples/async_io_watch_line_value.rs
@@ -22,7 +22,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     );
     async_io::block_on(async {
         while let Ok(event) = req.read_edge_event().await {
-            println!("{:?}", event);
+            println!("{event:?}");
         }
     });
     Ok(())