+++ cli/src/common.rs
@@ -390,20 +390,19 @@ pub fn format_chip_name(n: &str) -> &str {
 }
 
 pub fn format_time(evtime: u64, timefmt: &TimeFmt) -> String {
-    use chrono::{Local, NaiveDateTime, TimeZone, Utc};
+    use chrono::{Local, TimeZone, Utc};
 
     let ts_sec = (evtime / 1000000000) as i64;
     let ts_nsec = (evtime % 1000000000) as u32;
     match timefmt {
         TimeFmt::Seconds => format!("{}.{:09}", ts_sec, ts_nsec),
         TimeFmt::Localtime => {
-            let t = Local
-                .from_utc_datetime(&NaiveDateTime::from_timestamp_opt(ts_sec, ts_nsec).unwrap());
+            let t = Local.timestamp_opt(ts_sec, ts_nsec).unwrap();
             format!("{}", t.format("%FT%T%.9f"))
         }
         TimeFmt::Utc => {
             let t =
-                Utc.from_utc_datetime(&NaiveDateTime::from_timestamp_opt(ts_sec, ts_nsec).unwrap());
+                Utc.timestamp_opt(ts_sec, ts_nsec).unwrap();
             format!("{}", t.format("%FT%T%.9fZ"))
         }
     }